add logging for browserviews
